requirements.txt
appnope==0.1.4
asttokens==3.0.0
chardet==5.2.0
click==8.2.1
colorama==0.4.6
comm==0.2.3
coverage==7.10.3
debugpy==1.8.16
decorator==5.2.1
diff_cover==9.6.0
executing==2.2.0
flake8==7.3.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mirakuru==2.6.1
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
port-for==0.7.4
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
pytest-postgresql==7.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
regex==2025.7.34
setuptools==78.1.0
six==1.17.0
SQLAlchemy==2.0.43
sqlfluff==3.4.2
stack-data==0.6.3
tblib==3.1.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
wheel==0.45.1
